revertccl: move revert tenant helper
Release note: none.
Epic: none.
