+++ pkg/ccl/revertccl/BUILD.bazel
@@ -2,19 +2,30 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "revertccl",
-    srcs = ["revert.go"],
+    srcs = [
+        "revert.go",
+        "revert_tenant.go",
+    ],
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/revertccl",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/keys",
         "//pkg/kv",
         "//pkg/kv/kvpb",
+        "//pkg/kv/kvserver/protectedts/ptpb",
+        "//pkg/multitenant/mtinfopb",
         "//pkg/roachpb",
         "//pkg/settings",
         "//pkg/sql",
+        "//pkg/sql/clusterunique",
+        "//pkg/sql/isql",
+        "//pkg/sql/sem/eval",
+        "//pkg/sql/sessionprotectedts",
         "//pkg/util/hlc",
         "//pkg/util/log",
         "//pkg/util/syncutil",
         "//pkg/util/tracing",
+        "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
         "@org_golang_x_sync//errgroup",
     ],
