+++ pkg/ccl/streamingccl/streamingest/stream_ingest_manager.go
@@ -16,9 +16,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/ccl/utilccl"
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
-	"github.com/cockroachdb/cockroach/pkg/keys"
-	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/protectedts/ptpb"
-	"github.com/cockroachdb/cockroach/pkg/multitenant/mtinfopb"
 	"github.com/cockroachdb/cockroach/pkg/repstream"
 	"github.com/cockroachdb/cockroach/pkg/repstream/streampb"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
@@ -29,11 +26,8 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/eval"
-	"github.com/cockroachdb/cockroach/pkg/sql/sessionprotectedts"
 	"github.com/cockroachdb/cockroach/pkg/sql/syntheticprivilege"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
-	"github.com/cockroachdb/cockroach/pkg/util/log"
-	"github.com/cockroachdb/cockroach/pkg/util/uuid"
 	"github.com/cockroachdb/errors"
 )
 
@@ -55,115 +49,7 @@ func (r *streamIngestManagerImpl) GetReplicationStatsAndStatus(
 func (r *streamIngestManagerImpl) RevertTenantToTimestamp(
 	ctx context.Context, tenantName roachpb.TenantName, revertTo hlc.Timestamp,
 ) error {
-	return revertTenantToTimestamp(ctx, r.evalCtx, tenantName, revertTo, r.sessionID)
-}
-
-func revertTenantToTimestamp(
-	ctx context.Context, evalCtx *eval.Context, tenantName roachpb.TenantName, revertTo hlc.Timestamp, sessionID clusterunique.ID,
-) error {
-	execCfg := evalCtx.Planner.ExecutorConfig().(*sql.ExecutorConfig)
-
-	// These vars are set in Txn below. This transaction checks
-	// the service state of the tenant record, moves the tenant's
-	// data state to ADD, and installs a PTS for the revert
-	// timestamp.
-	//
-	// NB: We do this using a different txn since we want to be
-	// able to commit the state change during the
-	// non-transactional RevertSpans below.
-	var (
-		originalDataState mtinfopb.TenantDataState
-		tenantID          roachpb.TenantID
-		ptsCleanup        func()
-	)
-	if err := execCfg.InternalDB.Txn(ctx, func(ctx context.Context, txn isql.Txn) error {
-		tenantRecord, err := sql.GetTenantRecordByName(ctx, execCfg.Settings, txn, tenantName)
-		if err != nil {
-			return err
-		}
-		tenantID, err = roachpb.MakeTenantID(tenantRecord.ID)
-		if err != nil {
-			return err
-		}
-
-		if tenantID.Equal(roachpb.SystemTenantID) {
-			return errors.New("cannot revert the system tenant")
-		}
-
-		if tenantRecord.ServiceMode != mtinfopb.ServiceModeNone {
-			return errors.Newf("cannot revert tenant %q (%d) in service mode %s; service mode must be %s",
-				tenantRecord.Name,
-				tenantRecord.ID,
-				tenantRecord.ServiceMode,
-				mtinfopb.ServiceModeNone,
-			)
-		}
-
-		originalDataState = tenantRecord.DataState
-
-		ptsCleanup, err = protectTenantSpanWithSession(ctx, txn, execCfg, tenantID, sessionID, revertTo)
-		if err != nil {
-			return errors.Wrap(err, "protecting revert timestamp")
-		}
-
-		// Set the data state to Add during the destructive operation.
-		tenantRecord.LastRevertTenantTimestamp = revertTo
-		tenantRecord.DataState = mtinfopb.DataStateAdd
-		return sql.UpdateTenantRecord(ctx, execCfg.Settings, txn, tenantRecord)
-	}); err != nil {
-		return err
-	}
-	defer ptsCleanup()
-
-	spanToRevert := keys.MakeTenantSpan(tenantID)
-	if err := revertccl.RevertSpansFanout(ctx, execCfg.DB, evalCtx.JobExecContext.(sql.JobExecContext),
-		[]roachpb.Span{spanToRevert},
-		revertTo,
-		false, /* ignoreGCThreshold */
-		int64(sql.RevertTableDefaultBatchSize),
-		nil /* onCompletedCallback */); err != nil {
-		return err
-	}
-
-	return execCfg.InternalDB.Txn(ctx, func(ctx context.Context, txn isql.Txn) error {
-		tenantRecord, err := sql.GetTenantRecordByName(ctx, execCfg.Settings, txn, tenantName)
-		if err != nil {
-			return err
-		}
-		tenantRecord.DataState = originalDataState
-		return sql.UpdateTenantRecord(ctx, execCfg.Settings, txn, tenantRecord)
-	})
-}
-
-func protectTenantSpanWithSession(
-	ctx context.Context,
-	txn isql.Txn,
-	execCfg *sql.ExecutorConfig,
-	tenantID roachpb.TenantID,
-	sessionID clusterunique.ID,
-	timestamp hlc.Timestamp,
-) (func(), error) {
-	ptsRecordID := uuid.MakeV4()
-	ptsRecord := sessionprotectedts.MakeRecord(
-		ptsRecordID,
-		[]byte(sessionID.String()),
-		timestamp,
-		ptpb.MakeTenantsTarget([]roachpb.TenantID{tenantID}),
-	)
-	log.Infof(ctx, "protecting timestamp: %#+v", ptsRecord)
-	pts := execCfg.ProtectedTimestampProvider.WithTxn(txn)
-	if err := pts.Protect(ctx, ptsRecord); err != nil {
-		return nil, err
-	}
-	releasePTS := func() {
-		if err := execCfg.InternalDB.Txn(ctx, func(ctx context.Context, txn isql.Txn) error {
-			pts := execCfg.ProtectedTimestampProvider.WithTxn(txn)
-			return pts.Release(ctx, ptsRecordID)
-		}); err != nil {
-			log.Warningf(ctx, "failed to release protected timestamp %s: %v", ptsRecordID, err)
-		}
-	}
-	return releasePTS, nil
+	return revertccl.RevertTenantToTimestamp(ctx, r.evalCtx, tenantName, revertTo, r.sessionID)
 }
 
 func newStreamIngestManagerWithPrivilegesCheck(
